requirements.txt
streamlit
pandas
torch
transformers
pinecone-client
nltk
moviepy
wave
PyPDF2
sentence_transformers
clip
opencv-python-headless
git+https://github.com/openai/CLIP.git
